GameTimeに関係する整数型の警告に対応

+++ src/qt/score_board_painter_rcss.cpp
@@ -162,7 +162,7 @@ ScoreBoardPainterRCSS::draw( QPainter & painter )
                   right_team.score(),
                   ( right_team.name().empty() || opt.anonymousMode() ) ? "" : right_team.name().c_str(),
                   mode_str.c_str(),
-                  view->time().cycle() );
+                  static_cast< int >( view->time().cycle() ) );
         text = QString::fromLatin1( buf );
     }
     else
